+++ .gitignore
@@ -9,3 +9,4 @@ run_upscaller.bat
 index.html
 gui_test.py
 build/
+results/
