+++ .vscode/tasks.json
@@ -14,7 +14,7 @@
                 "panel": "new"
             },
             "options": {
-                "cwd": "${workspaceFolder}"
+                "cwd": "${workspaceFolder}/backend"
             },
             "problemMatcher": [],
             "isBackground": true,
@@ -35,7 +35,7 @@
                 "panel": "new"
             },
             "options": {
-                "cwd": "${workspaceFolder}/gui-nextjs"
+                "cwd": "${workspaceFolder}/frontend"
             },
             "problemMatcher": [],
             "dependsOn": "Start FastAPI Server"
@@ -71,24 +71,7 @@
                 "panel": "new"
             },
             "options": {
-                "cwd": "${workspaceFolder}/gui-nextjs"
-            },
-            "problemMatcher": []
-        },
-        {
-            "label": "Start Streamlit GUI",
-            "type": "shell",
-            "command": "python",
-            "args": ["-m", "streamlit", "run", "upscaler_gui.py"],
-            "group": "build",
-            "presentation": {
-                "echo": true,
-                "reveal": "always",
-                "focus": false,
-                "panel": "new"
-            },
-            "options": {
-                "cwd": "${workspaceFolder}"
+                "cwd": "${workspaceFolder}/frontend"
             },
             "problemMatcher": []
         }
