+++ Notes.txt
@@ -1,6 +1,7 @@
-- Downloaded files model weights are way too small to be complete. Look into using curl or similar for the larger weights
-	see this (https://stackoverflow.com/questions/25010369/wget-curl-large-file-from-google-drive/39225039#39225039)
-
-- jpg files can actually reduce in size after the upscaling due to the similarity of pixels
 - develop someway to introduce more grain and texture to upscalled image
 
+Backend todo:
+Containerize app
+add to azure
+add secrets if any
+add github actions workflow for automatic deployment
