+++ backend/api_server.py
@@ -7,7 +7,6 @@ import asyncio
 import json
 import logging
 import os
-import tempfile
 import uuid
 from io import BytesIO
 from pathlib import Path
@@ -22,13 +21,13 @@ from PIL import Image
 import numpy as np
 
 from upscale import ModelManager
-from util_file import process_byte_input
+from util_file import process_byte_input, generate_filename
 
 # Configure logging
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-app = FastAPI(title="Image Upscaler API", version="1.0.0")
+app = FastAPI(title="Image Upscaler API", version="0.1.0")
 
 # Configure CORS for Next.js frontend
 app.add_middleware(
@@ -90,37 +89,7 @@ async def send_progress_update(job_id: str, progress: float, message: str):
         except Exception as e:
             logger.error(f"Error sending WebSocket message: {e}")
 
-# Generate output filename        
-def generate_filename(filename, scale_list, resample_mode):
-    
-    extension = filename.split('.')[-1]
-    name_without_ext = '.'.join(filename.split('.')[:-1])
-    
-    factors_str = ' '.join([f"x{scale}" for scale in scale_list])
-    
-    # Check if filename ends with factors in parentheses like "(x2)" or "(x2 x4)"
-    if name_without_ext.endswith(')') and '(' in name_without_ext:
-        # Find the last opening parenthesis
-        last_open_paren = name_without_ext.rfind('(')
-        
-        # Extract what's inside the parentheses at the end
-        content_in_parens = name_without_ext[last_open_paren+1:-1]
-        
-        # Check if it contains factors (starts with 'x' and contains digits)
-        if content_in_parens and all(part.startswith('x') and part[1:].isdigit() for part in content_in_parens.split()):
-            # It's factors at the end, combine them
-            name_before_factors = name_without_ext[:last_open_paren]
-            combined_factors = f"{content_in_parens} {factors_str}"
-            new_filename = f"{name_before_factors} {resample_mode} ({combined_factors}).{extension}"
-        else:
-            # Not factors, treat as normal filename
-            new_filename = f"{name_without_ext} {resample_mode} ({factors_str}).{extension}"
-    else:
-        # No parentheses at end, add factors normally
-        new_filename = f"{name_without_ext} {resample_mode} ({factors_str}).{extension}"
-        
-    return new_filename
-    
+
 
 @app.get("/")
 async def root():
