+++ backend/util_file.py
@@ -28,19 +28,13 @@ async def process_byte_input(uploaded_file):
     else:
         raise ValueError("No file uploaded or file is empty.")
 
-
-def build_filename(original_filename, selected_scales=["2"]):
-    """ 
-        creates a new filename for the downloaded image that has the factor appended e.g image (x2).png 
-        Used for CLI
-    """
-    # Create new filename with factors appended
+# Generate output filename        
+def generate_filename(filename, scale_list, resample_mode):
     
-    extension = original_filename.split('.')[-1]
-    name_without_ext = '.'.join(original_filename.split('.')[:-1])
-
-    # Format new factors
-    factors_str = ' '.join([f"x{scale}" for scale in selected_scales])
+    extension = filename.split('.')[-1]
+    name_without_ext = '.'.join(filename.split('.')[:-1])
+    
+    factors_str = ' '.join([f"x{scale}" for scale in scale_list])
     
     # Check if filename ends with factors in parentheses like "(x2)" or "(x2 x4)"
     if name_without_ext.endswith(')') and '(' in name_without_ext:
@@ -55,37 +49,24 @@ def build_filename(original_filename, selected_scales=["2"]):
             # It's factors at the end, combine them
             name_before_factors = name_without_ext[:last_open_paren]
             combined_factors = f"{content_in_parens} {factors_str}"
-            new_filename = f"{name_before_factors}({combined_factors}).{extension}"
+            new_filename = f"{name_before_factors} {resample_mode} ({combined_factors}).{extension}"
         else:
             # Not factors, treat as normal filename
-            new_filename = f"{name_without_ext} ({factors_str}).{extension}"
+            new_filename = f"{name_without_ext} {resample_mode} ({factors_str}).{extension}"
     else:
         # No parentheses at end, add factors normally
-        new_filename = f"{name_without_ext} ({factors_str}).{extension}"
+        new_filename = f"{name_without_ext} {resample_mode} ({factors_str}).{extension}"
         
-
     return new_filename
 
 
-# TODO: allow selection of multiple images
-def file_select_dialog():
-    import tkinter as tk
-    from tkinter import filedialog
-
-    root = tk.Tk()
-    root.withdraw()
-
-    file_path = filedialog.askopenfilename()
-    return file_path
-
 def import_model():
-
     # Fall back to local directory import
     realesrgan_path = os.path.join(os.path.dirname(__file__), 'Real-ESRGAN')
     if realesrgan_path not in sys.path:
         sys.path.insert(0, realesrgan_path)
     try:
-        import RealESRGAN
+        import RealESRGAN # type: ignore
         print("RealESRGAN imported from local directory.")
     except ImportError:
         raise ImportError("RealESRGAN not found. Please install it via pip or place it in the 'Real-ESRGAN' directory.")
@@ -103,6 +84,22 @@ def download_gdrive_weights(URL, fileinx, dest):
     import sys
     import requests
     
+    def get_confirm_token(res):
+        for key, value in res.cookies.items():
+            if key.startswith("download_warning"):
+                return value
+
+        return None
+    def save_res_data(res, dest):
+        CHUNK_SIZE = 32768
+
+        with open(dest, "wb") as f:
+            for chunk in res.iter_content(CHUNK_SIZE):
+                if chunk:  # filter out keep-alive new chunks
+                    f.write(chunk)
+
+    
+    
     session = requests.Session()
     response = session.get(URL, params={"id": fileinx}, stream=True)
     token = get_confirm_token(response)
@@ -113,19 +110,3 @@ def download_gdrive_weights(URL, fileinx, dest):
         
     save_res_data(response, dest)
 
-
-def get_confirm_token(res):
-    for key, value in res.cookies.items():
-        if key.startswith("download_warning"):
-            return value
-
-    return None
-
-
-def save_res_data(res, dest):
-    CHUNK_SIZE = 32768
-
-    with open(dest, "wb") as f:
-        for chunk in res.iter_content(CHUNK_SIZE):
-            if chunk:  # filter out keep-alive new chunks
-                f.write(chunk)
