+++ frontend/src/components/ImageUpscaler.tsx
