--- gui-nextjs/package.json
@@ -1,77 +1,0 @@
-{
-  "name": "gui-nextjs",
-  "version": "0.1.0",
-  "private": true,
-  "main": "public/electron.js",
-  "homepage": "./",
-  "scripts": {
-    "dev": "next dev --turbopack",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "electron": "cross-env NODE_ENV=development electron .",
-    "electron-dev": "concurrently \"pnpm run dev\" \"wait-on http://localhost:3000 && cross-env NODE_ENV=development electron .\"",
-    "electron-pack": "pnpm run build && electron-builder",
-    "preelectron-pack": "pnpm run build"
-  },
-  "dependencies": {
-    "@radix-ui/react-checkbox": "^1.3.2",
-    "@radix-ui/react-dialog": "^1.1.14",
-    "@radix-ui/react-label": "^2.1.7",
-    "@radix-ui/react-progress": "^1.1.7",
-    "@radix-ui/react-select": "^2.2.5",
-    "@radix-ui/react-slider": "^1.3.5",
-    "@radix-ui/react-slot": "^1.2.3",
-    "@radix-ui/react-toast": "^1.2.14",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
-    "lucide-react": "^0.525.0",
-    "next": "15.4.2",
-    "react": "19.1.0",
-    "react-compare-slider": "^3.1.0",
-    "react-dom": "19.1.0",
-    "react-tooltip": "^5.29.1",
-    "tailwind-merge": "^3.3.1"
-  },
-  "devDependencies": {
-    "@eslint/eslintrc": "^3",
-    "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "concurrently": "^9.2.0",
-    "cross-env": "^7.0.3",
-    "electron": "^37.2.3",
-    "electron-builder": "^26.0.12",
-    "eslint": "^9",
-    "eslint-config-next": "15.4.2",
-    "tailwindcss": "^4",
-    "typescript": "^5",
-    "wait-on": "^8.0.3"
-  },
-  "build": {
-    "appId": "com.yourcompany.image-upscaler",
-    "productName": "Image Upscaler",
-    "directories": {
-      "output": "dist"
-    },
-    "files": [
-      "out/**/*",
-      "public/electron.js",
-      "public/preload.js",
-      "node_modules/**/*"
-    ],
-    "win": {
-      "target": "nsis",
-      "icon": "public/icon.ico"
-    },
-    "mac": {
-      "target": "dmg",
-      "icon": "public/icon.icns"
-    },
-    "linux": {
-      "target": "AppImage",
-      "icon": "public/icon.png"
-    }
-  }
-}
